Consolidate font colours in inverse style
A complication with removing the icon font and replacing it with an svg/png
served by the build service is that we have to add styling to the :after
psuedo selector of the button to force the browser to download the icon of the
hover state colour so there is no FOIC.

For this to work, buttons can only use two different font colours (as there
can only be one download on the :after selector). Inverse has two font colours,
white and ft-pink, where ft-pink is for pressed/selected and white is for
active. This PR makes the white ft-pink instead.

The next button styles are due some sprucing anyway, so this change is a
pragmatic stop-gap ahead of some more major refactoring that we should do when
the new button designs in next land in Origami.