Support custom themes with hex values and colour names.
We configure branded components like so:

```scss
@if oBrandGetCurrentBrand() == 'master' {
	@include oBrandDefine('o-example', 'master', (
		'variables': (
			'background': oColorsByName('teal')
			'inverse': (
				'background': oColorsByName('slate')
			)
		),
		'supports-variants': ()
	));
}
```

We can then get values for the current brand and variant using
a function like `oExampleGet`.

```scss
.inverse {
    background: oExampleGet('background', $from: 'inverse');
}
```

We can't use this configuration to generate buttons unless
o-buttons allows themes to be added with hex values.

This commit adds support for defining button themes with hex
values as well as colour names.

This way:
```scss
@include output($selector: false)  {
    @include oButtonsAddTheme(
        $name: 'example',
        $opts: (
            'color': 'teal-90',
            'context': 'slate'
        ),
        $types: ('primary', 'secondary')
    );
};
```

Or this way:
```scss
@include oButtonsAddTheme(
    $name: 'test',
    $opts: (
        'color': oColorsByName('teal-90'),
        'context': oColorsByName('slate')
    ),
    $types: ('primary', 'secondary'),
);
```

If the hex is not in the o-colors colour palette however an
error is thrown.