Make the active state take precedence over the hover state.
Our buttons do not give any feedback when clicked. That’s because
the hover state takes precedence over the active state, so the
active state isn’t shown until the cursor is moved away. The lack
of feedback is a pretty bad experience when using buttons to
interact with a page.

For an intimidate improvement, make the active state take precedence
over the hover state so we give that immediate feedback.

The drawback here is an active button then has no hover state.
We need to work with product design on what an active + hover state
looks like for primary and secondary buttons to improve that.