Remove `&[disabled]` from hover/focus state selectors.
Hover/Focus does not apply to disabled buttons so there is no
need to style hover/focus states when the button is  _not_
disabled.

Other things in this messy commit:
- Add hover icon to `before` pseudo element to flash as icon downloads
on hover (we already did this for the focus state with an `after`
pseudo element)
- Brings back conditional logic to only output icons of different
colours per button state if required. To support IE we can colour
masks with a new image request which is expensive, and good to
avoid where we can.