Remove demo JavaScript
The demo JavaScript toggles the `aria-selected` state on click.
This is a way to show the selected state in all button demos
but could be confusing to users, especially for patterns such
as pagination which involve multiple buttons and the selected state.
https://financialtimes.slack.com/archives/C02FU5ARJ/p1597227951166400

Instead, only show the selected state in relevant demos such as
the pagination demo.